Attempt to fix the error reported by the alpha tinderbox. A pointer was being cast to an integer as part of a hash function, so just add an intptr_t cast to silence the warning.
Former-commit-id: 90c1d9fb8bae944e26fb68034233813909e3a2c9

+++ sys/fs/coda/coda_namecache.h
@@ -86,7 +86,7 @@ parameter_list|,
 name|cp
 parameter_list|)
 define|\
-value|((name[0] + name[namelen-1] + namelen + (int)(cp))& (coda_nc_hashsize-1))
+value|((name[0] + name[namelen-1] + namelen + (int)(intptr_t)(cp))& (coda_nc_hashsize-1))
 end_define
 
 begin_else
@@ -106,7 +106,7 @@ parameter_list|,
 name|cp
 parameter_list|)
 define|\
-value|((name[0] + (name[namelen-1]<<4) + namelen + (((int)cp)>>8))& (coda_nc_hashsize-1))
+value|((name[0] + (name[namelen-1]<<4) + namelen + (((int)(intptr_t)cp)>>8))& (coda_nc_hashsize-1))
 end_define
 
 begin_endif
